added excellence and trivial autocomplete messages. Now Just the first Part has to be added to the .txt file. Also renamed the quotes.py file and merged the excellenz.py file into the new UserGeneratedContentFile
