+++ UserGeneratedContent/excellence_noAutocompletion.txt
@@ -1,0 +1,2 @@
+151 Jahre Exzellenz.
+Die Vorlesungssäle im MW sind sehr exzellent.
