+++ UserGeneratedContent/quotes.txt
