+++ UserGeneratedContent/trivial_autocomplete.txt
@@ -1,0 +1,7 @@
+Die TUM
+PgDP
+EiDI
+DS
+ERA
+Die Mensa
+Die Bockabfertigung in der Mensa
