--- exzellenz.txt
@@ -1,14 +1,0 @@
-151 Jahre Exzellenz.
-Die TUM ist sehr exzellent.
-PgDP ist sehr exzellent.
-EiDI ist sehr exzellent.
-DS ist sehr exzellent.
-ERA ist sehr exzellent.
-Die Mensa ist sehr exzellent.
-Die Bockabfertigung in der Mensa ist sehr exzellent.
-Die Vorlesungssäle im MW sind sehr exzellent.
-DS ist mir zu trivial.
-PgDP ist mir zu trivial.
-EiDI ist mir zu trivial.
-DS ist mir zu trivial.
-ERA ist mir zu trivial.
